requirements.txt
python-dotenv>=1.0.0
google-genai
numpy
sounddevice
pyaudio
google-cloud-speech
google.cloud.aiplatform 
vertexai
scikit-learn
pypdf2
pandas
PyMuPDF 
Pillow
faiss-cpu
pyarrow